fix warning of type hint

+++ conoha_client/features/image/domain/image.py
@@ -83,7 +83,7 @@ class Image(BaseModel, frozen=True):
         return v
 
     @field_serializer("sizeGB")
-    def _serialize(self, v: int) -> int:
+    def _serialize(self, v: int) -> float:
         """B to GB."""
         return v / pow(1024, 3)
